factor dom-particle into two classes
